ci: Build Xserver from git
This is to get support for the “-terminate delay” option in Xwayland
with libXfixes 6.0.

libXfixes 6.0 provides the new ClientDisconnectMode that clients can
use to declare themselves as to-be-terminated.

This can be used to let Xwayland terminate itself automatically when no
other (“non-terminatable”) X11 client is left.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/1794>